+++ services/preview/purview/mgmt/2020-12-01-preview/purview/accounts.go
@@ -149,6 +149,7 @@ func (client AccountsClient) CreateOrUpdatePreparer(ctx context.Context, resourc
 		"api-version": APIVersion,
 	}
 
+	account.Sku = nil
 	preparer := autorest.CreatePreparer(
 		autorest.AsContentType("application/json; charset=utf-8"),
 		autorest.AsPut(),